파일 업로드 컴포넌트 구조 분리 및 코드 재활용
