+++ src/ShortText/Upload/MediaUpload/components/QRCode.tsx
