+++ src/index.ts
@@ -15,8 +15,8 @@ export { default as ScaleSelection } from './ScaleSelection/ScaleSelection';
 export { default as Chart } from './Chart/Chart';
 export { default as CheckBoxItem } from './CheckBoxItem/CheckBoxItem';
 export { default as Tag } from './Tag/Tag';
-export { default as FileUpload } from './ShortText/FileUpload/FileUpload';
-export { default as ImgVideo } from './ShortText/MediaUpload/ImgVideoUpload';
+export { default as FileUpload } from './ShortText/Upload/FileUpload/FileUpload';
+export { default as ImgVideo } from './ShortText/Upload/MediaUpload/ImgVideoUpload';
 export { default as IEWarning } from './IEWarning/IEWarning';
 export { default as ScrollTop } from './ScrollTop/ScrollTop';
 export { default as ScrollDown } from './ScrollDown/ScrollDown';
